\name{polyads}
\docType{data}
\alias{polyads}
\title{Polyadic data example}

\description{
This data set contains a data frame with 10 triads. The 10 first rows correspond to the 1st generation members, the next 10 to the second generation members and the last 10 to the third generation members.}

\details{The first column is the generation variable Gen. The sequences of length 10 are in columns 2 to 11 named X1 to X10.
}

\usage{data(polyads)}
\format{A data frame with 30 rows and 11 columns.}
\keyword{Datasets}
\author{Gilbert Ritschard}
