\name{seqLRT}
\alias{seqLRT}
\alias{seqBIC}

\title{Likelihood ratio test and BIC for Comparing Sequence Data}

\description{The function \code{seqLRT} computes the likelihood ratio test (LRT) and BIC criteria for comparing
each pair of member sets in the two groups of sequences generated by \var{TraMineR}. The function \code{seqBIC} is an alias that returns only the Bayesian Information Criterion (BIC).
}

\usage{
seqLRT(seq1, seq2, s=100, sd=36963, stat="LRT",
    method=method, squared=FALSE, ...)

seqBIC(seq1, seq2, s=100, sd=36963, method=method, squared=FALSE, ...)
}

\arguments{
\item{seq1}{A list of the first set of input sequences. List is generated by R's
  \code{list()} function. E.g., list(cohort1.seq,cohort2.seq,cohort3.seq) where *.seq is
  generated by \code{\link[TraMineR]{seqdef}}.
}

\item{seq2}{A list of the second set of input sequences. When a group of sequences
  has more states, put that group in seq1 to facilitate computation.
}

\item{s}{Integer. Default 100. The number of randomly chosen sequences with replacement.
}

\item{sd}{Integer. Default 36963. Using the same seed number guarantees the same results
  each time.
}
\item{stat}{String. The requested statistics. One of \code{"LRT"}, \code{"BIC"}, or \code{"all"} }
\item{squared}{Logical. Should distances be squared?
}
\item{method}{String. Method for computing sequence distances. See documentation for \code{\link[TraMineR]{seqdist}}. Additional arguments may be required depending on the method chosen.
}
\item{...}{Additional parameters passed to \code{\link[TraMineR]{seqdist}}.
}

}

\value{The function \code{seqLRT} with the default \code{"LRT"} stat value value outputs two variables, \var{LRT} and \var{p.LRT}.

\item{LRT}{This is the likelihood ratio test statistic for comparing the two groups.
}
\item{p.LRT}{This is the upper tail probability associated with the LRT.
}
The function \code{seqBIC} (and \code{seqLRT} with the \code{"BIC"} stat value) outputs two variables, \var{BIC} and \var{BF}.

\item{BIC}{This is the difference between two BICs for comparing the two groups.
}
\item{BF}{This is the Bayes factor associated with the BIC difference.
}
}

\examples{
## biofam data set
data(biofam)
summary(biofam[,1:9])
## We use only a sample of 300 cases
set.seed(10)
biofam <- biofam[sample(nrow(biofam),300),]
biofam.lab <- c("Parent", "Left", "Married", "Left+Marr",
                "Child", "Left+Child", "Left+Marr+Child", "Divorced")
biofam.seq <- seqdef(biofam, 10:25, labels=biofam.lab)

## Defining the grouping variable
lang <- as.vector(biofam[["plingu02"]])
lang[is.na(lang)] <- "unknown"
lang <- factor(lang)

## Chronogram by language group
seqdplot(biofam.seq, group=lang)

## Extracting the sequence subsets by language
lev <- levels(lang)
l <- length(lev)
seq.list <- list()
for (i in 1:l){
  seq.list[[i]] <- biofam.seq[lang==lev[i],]
}

seqLRT(list(seq.list[[1]]),list(seq.list[[2]]), stat="all", method="OM", sm="CONSTANT")
seqBIC(list(seq.list[[1]]),list(seq.list[[2]]), method="HAM")
seqLRT(list(seq.list[[1]],seq.list[[1]]),list(seq.list[[2]],seq.list[[3]]), s=120, method="HAM")

}

\references{
Tim F. Liao & Anette E. Fasang. Forthcoming. "Comparing Groups of Life Course Sequences Using the Bayesian Information Criterion and the Likelihood Ratio Test.” \emph{Sociological Methodology} xx:xxx–xxx.
}
