\name{seqLRT}
\alias{seqLRT}
\alias{seqBIC}

\title{Likelihood ratio test and BIC for Comparing Sequence Data}

\description{The function \code{seqLRT} computes the likelihood ratio test (LRT) and BIC criteria for comparing
each pair of member sets in the two groups of sequences generated by \var{TraMineR}. The function \code{seqBIC} is an alias that returns only the Bayesian Information Criterion (BIC).
}

\usage{
seqLRT(seqdata, seqdata2=NULL, group=NULL, subgroup=NULL, 
    s=100, seed=36963, stat="LRT",
    squared=FALSE, method=method, ...)

seqBIC(seqdata, seqdata2=NULL, group=NULL, subgroup=NULL, 
    s=100, seed=36963, squared=FALSE, method=method, ...)
}

\arguments{
\item{seqdata}{Either a state sequence object (\code{stslist} created with \code{\link[TraMineR]{seqdef}}) or a list of state sequence objects, e.g., \code{list(cohort1.seq,cohort2.seq,cohort3.seq)}.
  %\code{list()} function. E.g., list(cohort1.seq,cohort2.seq,cohort3.seq) where *.seq is
  %generated by \code{\link[TraMineR]{seqdef}}.
}

\item{seqdata2}{Either a state sequence object (\code{stslist} or a list of state sequence objects. Must be \code{NULL} when \code{group} is not \code{NULL}. If not \code{NULL}, must be of same type than \code{seqdata}.
}
  \item{group}{Vector of length equal to number of sequences in \code{seqdata}. A dichotomous grouping variable. See details.
  }
  \item{subgroup}{Vector of length equal to number of sequences in \code{seqdata}. Variable defining the subgroups. See details.
  }

\item{s}{Integer. Default 100. The number of randomly chosen sequences with replacement.
}

\item{seed}{Integer. Default 36963. Using the same seed number guarantees the same results
  each time.
}
\item{stat}{String. The requested statistics. One of \code{"LRT"}, \code{"BIC"}, or \code{"all"} }
\item{squared}{Logical. Should squared distances be used?
}
\item{method}{String. Method for computing sequence distances. See documentation for \code{\link[TraMineR]{seqdist}}. Additional arguments may be required depending on the method chosen.
}
\item{...}{Additional parameters passed to \code{\link[TraMineR]{seqdist}}.
}

}

\details{
When \code{seqdata} and \code{seqdata2} are both provided, the LRT and BIC statistics are computed for comparing these two sets
}

\value{The function \code{seqLRT} with the default \code{"LRT"} stat value value outputs two variables, \var{LRT} and \var{p.LRT}.

\item{LRT}{This is the likelihood ratio test statistic for comparing the two groups.
}
\item{p.LRT}{This is the upper tail probability associated with the LRT.
}
The function \code{seqBIC} (and \code{seqLRT} with the \code{"BIC"} stat value) outputs two variables, \var{BIC} and \var{BF}.

\item{BIC}{This is the difference between two BICs for comparing the two groups.
}
\item{BF}{This is the Bayes factor associated with the BIC difference.
}
}

\examples{
## biofam data set
data(biofam)
## We use only a sample of 300 cases
set.seed(10)
biofam <- biofam[sample(nrow(biofam),300),]
biofam.lab <- c("Parent", "Left", "Married", "Left+Marr",
                "Child", "Left+Child", "Left+Marr+Child", "Divorced")
biofam.seq <- seqdef(biofam, 10:25, labels=biofam.lab)

## Defining the grouping variable
lang <- as.vector(biofam[["plingu02"]])
lang[is.na(lang)] <- "unknown"
lang <- factor(lang)

## Chronogram by language group
seqdplot(biofam.seq, group=lang)

## Extracting the sequence subsets by language
lev <- levels(lang)
l <- length(lev)
seq.list <- list()
for (i in 1:l){
  seq.list[[i]] <- biofam.seq[lang==lev[i],]
}

seqLRT(list(seq.list[[1]]),list(seq.list[[2]]), stat="all", method="OM", sm="CONSTANT")
seqBIC(biofam.seq, group=biofam$sex, method="HAM")
seqLRT(biofam.seq, group=biofam$sex, subgroup=lang, s=120, method="HAM")

}

\references{
Tim F. Liao & Anette E. Fasang. Forthcoming. "Comparing Groups of Life Course Sequences Using the Bayesian Information Criterion and the Likelihood Ratio Test.” \emph{Sociological Methodology} xx:xxx–xxx.
}
