\name{TraMineRextras-package}
\alias{TraMineRextras-package}
\alias{TraMineRextras}
\docType{package}
\title{ TraMineR Extension }
\description{
(Version: 0.7.7) Collection of ancillary functions and utilities to be used in conjunction with the 'TraMineR' package for sequence data exploration. Includes, among others, specific functions such as state survival plots, position-wise group-typical states, dynamic sequence indicators, and dissimilarities between event sequences. Also includes contributions by non-members of the TraMineR team such as the relative frequency plot and methods for polyadic data.
}

%\details{
%
%}

\author{ Gilbert Ritschard, Matthias Studer, Reto Buergin
}


\keyword{ package } 